Moves files from app/renderer/components into appropriate folder Renames ImportBrowserDataPanel into WidevinePanel in file widevinePanel.js
Auditors: @bsclifton

Test Plan:
- tests should be green
