+++ app/renderer/components/common/clipboardButton.js
