+++ app/renderer/components/preferences/payment/advancedSettings.js
@@ -11,7 +11,7 @@ const appConfig = require('../../../../../js/constants/appConfig')
 
 // components
 const BrowserButton = require('../../common/browserButton')
-const {SettingsList, SettingItem, SettingCheckbox} = require('../../settings')
+const {SettingsList, SettingItem, SettingCheckbox} = require('../../common/settings')
 const {SettingDropdown} = require('../../common/dropdown')
 const ImmutableComponent = require('../../immutableComponent')
