+++ js/components/main.js
@@ -23,7 +23,7 @@ const TabsToolbar = require('../../app/renderer/components/tabs/tabsToolbar')
 const FindBar = require('./findbar')
 const UpdateBar = require('./updateBar')
 const {NotificationBar} = require('./notificationBar')
-const DownloadsBar = require('../../app/renderer/components/downloadsBar')
+const DownloadsBar = require('../../app/renderer/components/download/downloadsBar')
 const SiteInfo = require('./siteInfo')
 const BraveryPanel = require('./braveryPanel')
 const ClearBrowsingDataPanel = require('../../app/renderer/components/clearBrowsingDataPanel')
